+++ README.md
@@ -21,7 +21,7 @@ poetry install
 4. Модуль generators. Одна функция-генератор фильтрации списка по валюте и две функции-генератора: одна собирает 
 описание операций по картам, другая генерирует номера карт от 0000 0000 0000 0000 до 9999 9999 9999 9999 (можно использовать 
 разные начальные и конечные значения)
-5. Модуль decorators. Содержит декоратор, записывающий логи выполнения функций в консоль или файл. Логи записываются в 
+5. Модуль decorators. Содержит декоратор, записывающий лог выполнения функций в консоль или файл. Логи записываются в 
 отдельный файл с указанным именем. При отсутствии имени файла логи будут выведены в консоль. 
 
 ## Тестирование
@@ -33,7 +33,7 @@ poetry install
 2. В модуле test_masks функции карты и счета тестируются поотдельности.
 3. В модуле test_processing тестируются функции фильтрации и сортировки списков по задаваемым параметрам.
 4. В модуле test_generators тестируются функции-генераторы.
-5. В модуле test_decorators идет проверка по предикатору.
+5. В модуле test_decorators идет проверка декоратора по предикату. Также идет проверка вывода в консоль.
 #### Результаты тестирования и покрытие кода находятся в `/htmlcov/index.html`
 
 ## Документация:
